requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
Jinja2==3.0.3
jio-health-toolkit-cache-python==0.2.4
jio-health-toolkit-hacks-python==0.0.1
jio-health-toolkit-logging==0.1.15
SQLAlchemy==1.3.15
PyMySQL==1.0.2
itsdangerous==2.0.1
qrcode==7.3.1
Pillow==9.0.1